third_party/blink/web_tests/external/wpt/css/css-grid/grid-lanes/tentative/parsing/flow-tolerance-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['flow-tolerance'] = "normal" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "infinite" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "initial" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "0" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "1px" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "calc(2em + 3ex)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "4%" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['flow-tolerance'] = "5vmin" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
